Move the accountant config to the system directory
